GH-55 - Implement dynamic relationships.
We had tests in place that allowed using `Map<String, Object>` as definition of a dynamic relationship.
The key of the map represents the type of the relationship in that case.

The commit adds to the logic of writing, reading and deleting relationships: A relationship is dynamic, if the underlying property is a map.
In such case, no type is used to write, read and delete relationships.
While in theory one could handle all queries without a type, that would be slower than
necessary (Cypher / database wise), so we handled that locally.

The mapping function needs to be adapted to correctly extract added information to the query result.
When it works on a system generated query, it looks for an artifical property, if it works on a query
defined by the user, it expect the relationship to be contained in the result and uses `type()` on
the client.