GH-54 - Derive default type for relationships if none is given.
This introduces a method similar to toUpperSnakeCase from OGM, to derive
a type for relationship when none is given.

The computation of the relationship description has been moved to the
Neo4jPersistentProperty, as the contract already has a method for
computing an association.

The contract can be fulfilled by extending from the associciation und
thus removing some clutter from the mapping context.