GH-25 - Introduce additional Neo4j template and reactive templates.
While being rather giantic, this commit doesn’t change too much:
Most of the methods from SimpleNeo4jRepository and it’s counterpart have been pushed down to a Neo4jTemplate and ReactiveNeo4jTemplate.

Those templates are implementations of an Operations interface, similar to what other SD projects have.

The biggest chain is actually the move of the Cypher generator to it’s own class and the removal of the relationship from the schema.
Those can be retrieved directly on the NodeDescription.

This gives us
* Templates
* Purely schema based access (only using package schema)
* And nice SimpleXXXNeo4jRepositories (the constructor doesn’t expose the kitchensink and the rest anymore) so that they can be neatly used as base for custom repositories.

This closes #25.