GH-112 - Add support for derived countBy and pageable methods.
This add support for derived countBy and pageable methods. It removes the need for having the result processor in concrete classes (part and string based version) and allows the imperative version to execute an additional count query for building the page.

This also addresses the fact that the indexes of bindable parameters are not necessarly dense, that is: They can contain special, non bindable parameters before and in between. This has been fixed with an index supplier.

This fixes #112.