Decouple default callbacks from repository configuration.
This commit introduces a pair of registrars for the default imperative and reactive callbacks (Id generation and auditing) so that a user might use the templates without enabling repositories. The registrar are meant to be public API, so that they are available in both our configuration support as well as in the Spring Boot starter (and maybe scenarios a user thinks off).