Introduce ReactiveNeo4jTest.
This test slice provides only reactive capatiblities and can be used in scenarios where `spring.data.neo4j.repositories.type` is set to `reactive` without the need to disable the configuration of `@Transactional`.

This slices checks for the Neo4j version before the test and thus fails without having at least Neo4j harness 4.0.0 on the classpath and indicates so with a decent error message.