Allow for spatial queries with Box and Polygon.
This brings support for the spatial queries within a box or a polygon.

We cannot yet transform those into native spatial queries, so we compute the bounding box of the box or the polygon and use that. In case of polygons, this might lead to „false positive“ depending on the orientation of the polygon.

This closes #191.