Exclude associations filled by constructor from property based handling.
In a scenario in which an entity needs both constructor and property population (for example having a non final, internally generated id and some fixed, unchangeable business values), associations filled in via a constructor should not be handled via the property handler.

In cases where an accessor exisits, this does duplicate work, in cases in which they are final, things will fail. This commit fixes this behaviour.