Add named paths and missing predicate functions.
This adds the ability to use named paths as defined in 3.1.4.8 of the docs.

Based on those (and other expressions), all list based predicates have been added as well (`any`, `all`, `single` and `none)`.
The existing `exists` predicate for properties and patterns have been moved from the `Conditions` to `Predicates` as well.

This closes #247.